integrate redux with react
