+++ src/App.js
@@ -1,13 +1,14 @@
 import React, { Component } from 'react';
+import Counter from './components/Counter';
 
 class App extends Component {
   render() {
     return (
       <div className="App" >
-        Counter Component here.
+        <Counter store={this.props.store} />
       </div>
     );
-  }
-}
+  };
+};
 
 export default App;
