+++ src/createStore.js
@@ -1,1 +1,21 @@
-
+import { render } from './index.js';
+ 
+export default function createStore(reducer) {
+  let state;
+ 
+  function dispatch(action) {
+    state = reducer(state, action)
+    console.log(`the state is ${state.count}`)
+    console.log(`the action is ${action.type}`)
+    render();
+  };
+ 
+  function getState() {
+    return state;
+  };
+ 
+  return {
+    dispatch,
+    getState
+  };
+};
