+++ src/reducers/changeCount.js
@@ -1,0 +1,9 @@
+export default function changeCount(state = {count: 0}, action) {
+  switch (action.type) {
+    case 'INCREASE_COUNT':
+      return { count: state.count + 1 };
+ 
+    default:
+      return state;
+  };
+};
